Split rust and C brotli decoders into separate submodules
brotli.rs becomes brotli/{mod,rust,c}.rs: mod.rs holds the module docs
and the shared decompress_woff2 entry point; rust.rs holds the
brotli-decompressor backend (Rebox/HeapAlloc + BrotliDecompressStream);
c.rs holds the brotlic backend. Feature selection moves to cfg-gated
`mod`/`use` statements in mod.rs instead of cfg attributes scattered
through one file.
